refactor: fetch async로 바꾸기
