+++ src/index.js
@@ -1,7 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import "./css/index.css";
-import App from "./App";
+import App from "./components/App";
 import reportWebVitals from "./env/reportWebVitals";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
